[fxfs] Fix a number of issues with deleting volumes
The main issue was that when replaying the MarkForDeletion mutation, we
would only actually apply it if `owner_bytes` had something present, but
that's not guaranteed when replaying the journal.

Included in this change:

 + I simplified the way we track volumes pending deletions.  This CL
   will now flush the device a bit more eagerly, but deleting volumes
   should be a rare event.

 + I split flushing into two halves.  I wanted to do this so I could
   test what happens between the transactions involved in flushing but
   ended up not needing it.

 + I removed the call to did_flush_device in the allocator because it
   was passing the wrong value for the offset, and it's easier if the
   allocator assumes the device has been flushed in `Allocator::open`.

 + Better checks on the byte counters.

I wasn't able to reproduce the *exact* log message that was seen in the
bug, but I believe the bugs fixed here probably could have caused what
was seen in the bug.  Regardless, if we encounter more issues, the
improved logging included in this change should give us more
information.

Bug: b/341928701
Change-Id: I2b5c297fd047aa8d42cd2eff1e4f167c3cf63738
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058212
Reviewed-by: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>