[rbe][link] Avoid unwanted option prefix matching
argparse.parse_known_args() aggressively tries to
match against unique prefixes of options (even with allow_abbrev=False),
which results in unwanted interpretation of '-r' as '-rtlib',
and '-other' as ('-o' 'ther').

To workaround this, we (under-the-hood) register options like '-rtlib'
and '-o' in their double-dash form like '--o' and '--rtlib', and rewrite
the args before passing to ArgumentParser.parse_known_args().

We can't apply this broadly because of the existence of fused-style
flags '-L' and '-W'.

Unit tests cover cases where potentially matched short forms
of options do *not* get matched, and end up in the set of uninterpreted
args after parse_known_args().

This keeps with the spirit of minimizing maintenance as
new unfamiliar flags get used in the build.

Fixed: 343808592
Change-Id: I551457833b58c9df6522911fe67ecc9f3b4bb822
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058358
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: David Fang <fangism@google.com>