[perfcompare] Add "--expected_sample_size" arg to provide a consistency check
This specifies the expected sample size (number of boots) for each
metric in each set of results.

This is intended to catch cases where an Infra build improperly runs a
test more than once, such as on a mix of inconsistent device types
(e.g. both NUC7 and NUC11, or both NUC and QEMU).

This argument will be passed by the Infra recipe, which will be done
in this change:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/1058660

This argument will be passed when displaying the table of performance
results for perfcompare and non-perfcompare builds.

Change the existing unit tests to use this argument so that it gets
exercised as much as possible by the unit tests.

Bug: 343997506
Run-All-Tests: True
Change-Id: I88794a19d09fdbe64940263e14f987d2a4d1b946
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058663
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>