[fxfs] ProfileState object is recreated on the volume for each use.
The `ProfileState` object is passed into the
`FxVolume.record_or_replay_profile()` call and dropped when done. This
is primarily to delegate which `ProfileState` implementation will be
created to the `VolumesDirectory` as the `FxVolume` itself does not know
if it is a blob volume or not.

Bug: b/337321099

Change-Id: I4bcb174c0ec6077e6d80c04ec681ea2719a1837a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049614
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>