[superproject] Roll antlion Fix WNM config for 5 GHz hostapd AP.
When configuring WNM features (e.g. BSS Transition Management) for
hostapd AP on 5 GHz, the WNM features were not being added to the
config. This was not noticed because the firmware used for current
Fullmac devices seems to be able to reassociate through alternate logic
when BSS Transition Management is not advertised. WNM features are not
in use yet on production devices, so this did not affect testing for
production devices.

This change corrects the error, and corrects an ordering issue in the
construction of the hostapd config. I manually verified via packet
capture that the new logic advertises WNM features correctly on both
2.4 and 5 GHz hostapd APs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1059255
Original-Revision: 1dbff7bb3b9a650a6d2ddb2910499bf1a952f21f
GitOrigin-RevId: f9e5579b94ce4ed88ad16907d049be696dada16d
Roller-URL: https://ci.chromium.org/b/8746031816848233633
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ib352d7c2a4d0692e1fd83311fe77ae90a9060919
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060136