[package-directory] Use VmoFile for the meta file
MetaAsFile serves the the package's hash as a file. By writing the hash
to a VMO, VmoFile can be used to the serve the hash instead. VmoFile
supports streams which are significantly faster than
fuchsia.io/File.Read for reading the file's contents.

The downside is that the meta file will now use a full page of memory
instead of 64 bytes while being served. This should not be an issue
because the meta file is typically only around for a very short period
of time.

Bug: b/327633753
Change-Id: I4912eea607c8a14695fcf766cfdc54bad04a1612
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009073
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>