[fidl] Optimize the client waker
On Fuchsia, when there are multiple wakers, we can read the channel when
we are woken and then only wake the task that needs to be woken rather
than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

This is a reland. I have rewritten how the client works when there is
only a single waker to avoid the issues that caused the reverts (namely
the raciness around `into_channel`).

Change-Id: Ibe7a7490c4e41dff32d9c64e93e130b6a4d5d13a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058952
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Chris Suter <csuter@google.com>