[display][edid] Use cpp20::span for Edid initialization.
This change replaces the Edid::Init() method with a raw pointer
and a uint16_t size variable with the Init() method using a
cpp20::span. This removes the usage of the uint16_t type as
container size, which is discouraged by the style guide.

Test: no behavior change.
Test: core.vim3 (with AML028-30MB-A1 HDMI display) booted.
Bug: 344042452
Change-Id: I319c9b5cf252eb31cbd92cdd08ca7065b2e0dc5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060379
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>