[fsverity] Stop verity benchmarks from being run twice
The fix here is simple but it's worth explaining the background.

Since they were added, the verity benchmarks were being run twice,
because of the gotcha described here:
https://fuchsia.dev/fuchsia-src/development/performance/faq
(under "Why does my performance test get run twice in Infra builds?")

Initially the second run was on QEMU, but later the second run was on
a real hardware device.

The change https://fuchsia-review.googlesource.com/c/fuchsia/+/1009460
("[terminal] Switch to Fxblob") switched the second run to be on real
hardware instead of QEMU, but neither arrangement was correct because
we don't want the test to be run a second time at all.

I noticed this by accident a few weeks ago.  I noticed that queries of
the culprit finder's performance results database were returning
multiple results per build for the metrics produced by this test.

It used to be that if a performance test were accidentally run as a
Fuchsia-side-only test a second time, its results would not get copied
into the Infra test shard's output, and so would not be published into
the culprit finder's results database.  So the gotcha above would not
cause the problem of getting duplicate results.  However, that was
apparently changed by this CL from November 2023:
https://fuchsia-review.googlesource.com/c/fuchsia/+/936577
"[testrunner] Collect test outputs from target-side tests."

Landing this CL is a dependency of enabling the following consistency
check, which should catch and prevent this problem in the future:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/1058660

Bug: 343997506
Change-Id: I26b90812e1039100b3b424cdf3a96094dce62291
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058721
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>