[ffx][playground] Run the tests for the playground tool
The tool is experimental but we should still run the tests even though we
haven't enabled it.

We also fix a circular reference issue in cf_fs that was making the
tests fail lsan. We do that in this CL so that the change is verified by
CQ.

Change-Id: Ibd12314c699ef1a465dd6fb6084808d32d174977
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058493
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>