[netstack3] Add support for ICMP6_FILTER
Allow raw IP sockets to set the ICMP6_FILTER socket option. This allows
the user to control which types of ICMP messages they would like to
receive on the socket. Note that this option is only allowed for IPv6
sockets created with protocol 58 (IPPROTO_ICMPV6).

The implementation in core is generic over IP version (e.g. ICMPv4
filters are allowed, despite not being exercised by bindings).

Notably the `FilterICMPPackets` gvisor syscall test is not passing; this
is because it relies on the netstack populating the checksum for sent
ICMPv6 messages, which Netstack3 does not yet do. A follow up CL should
allow this test to pass.

Fixed: 337816586
Change-Id: Ie2964ef0965dfb0fc1847b34e8424199f7397527
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058692
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>