[ffx] Daemonless `target wait` command.
This takes `ffx target wait` and makes the command daemonless. In order
for this to function the RCS knock timeout has been increased from 500ms
to 5 seconds, otherwise failures would be likely to occur given the
amount of time it takes for the ssh probe to make it through to the
device. This could have wider reaching consequences, but by default `ffx
target wait` has a timeout of 120 seconds still.

Bug: b/338436543
Change-Id: I0e5673bd8c4a5890d58a973487fb901fa49739a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042489
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>