[netstack3] Require DeviceIdContext in filtering
Currently a lot of code in the filtering crate is parameterized on a
generic device ID type that is constrained to implement
InterfaceProperties so it can be matched against. Given that in
practice, all this code is dealing with the same device ID type, and we
sometimes want to constrain that type further (e.g. with Debug),
centralize the type and the constraints on it by requiring that
FilterIpContext, FilterHandler, and FilterHandlerProvider all implement
DeviceIdContext (this was already true for NatContext). Further, require
that the associated DeviceId type implement InterfaceProperties.

Change-Id: I45bac78b5eb41f6bdc865968c3729eac10aa9fe8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060212
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>