[forensics] Add promise timeout utility
Adds a utility function that allows users to specify a timeout duration
for the promise passed to MakePromiseTimeout. Promises with a timeout
must have an error type of forensics::Error.

This code is not yet used, but will first be used to timeout if
acquiring a wake lease takes too long.

Bug: 333110044
Tested: `fx test promise_timeout_unittest`
Change-Id: Ib4f568162e3c0da268dbd6d0c09c44521ec64f12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060984
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>