[exceptions] Timeout wake lease acquisition
Timeout if acquiring a wake lease takes more than a specified amount of
time. It's important to release exceptions in a timely manner and
failing to acquire a lease within the specified time may indicate that
we'll never finish acquisition.

This class is not yet used, so the timeout value is not yet specified.

Bug: 333110044
Tested: `fx test wake_lease_unittest`
Change-Id: I2f8ec951cc6e4d2164b1fb3a568c529cc7f9d70b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061014
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>