[vfs][rust] Remove flags from TokenRegistry
The flags are unused.

Tokens are used to identify the destination directory of a mutating
operation. The connection that creates the token ensures that the
connection has the right to modify the directory. This avoids the
connection receiving the token from having to check the rights of the
connection that created the token.

Change-Id: I7df9276f8dc92b3accad42d7cde4f248b1f6ae68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058592
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>