[build] Create RBE 'modes' to simplify configuration
This creates named modes for RBE to operate in.  These modes then set
the default values for a scope of rbe settings.  The user can override
the settings in this scope.  This settings scope is then used to set
all the specific RBE flags.

Bug: b/344673187
Change-Id: Ic59de43bc7babca8ad382e6b233260d28a5c2c78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043387
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>