[fxfs] Add poison field to FxVolume to gather flake information.
When shutting down test fixtures in the fuchsia platform, the
verification that there are no other references to the volume
occasionally fails. There is no information then about where the extra
reference is being held. Now when the final drop eventually triggers we
will get a stack trace of where it was being held.

Bug: b/342837879
Change-Id: I5f60ffa16af5e80c8363122bf4049452063c285c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061532
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>