[idk] subbuild.py: Add --setup-only option.
Add an option to the script to allow skipping the Ninja build
part. This will be used in the future when setting up the sub-build
repositories in a Bazel repository rule (but performing the build
separately through a Bazel action).

+ Make `--prebuilt-host-toos-dir` optional.
+ Remove an obsolete assertion.

Change-Id: I41d17e0a183a1655ad94b66072143ee4720f227f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062134
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>