[netstack3] Introduce DatagramApi
This allows us to export less symbols in the crate's public API as well
as get rid of a lot of very verbose callsites.

This is mostly all code movement so that all the functions that are pub
fit in the same block. Care was taken to try and keep them in the same
order where possible, but given the spread in the file a lot of lines
are touched.

Change-Id: I51617f826f7c53d8d65413c7795838be4f6a737b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061639
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>