[netstack3] Remove vector allocation in addr check
The blanket impl for BaseIpTransportContext was allocating a vector to
satisfy the trait's signature, but we can do it without the allocation.

Change the shape to the `with_` form, so there's no impedance mismatch
with the traits this builds on and introduce a tailored iterator to
solve the problem for the blanket impl.

While I'm here, remove a now unnecessary lifetime bound on the devices
and addresses iterator in `IpDeviceContext`.

Change-Id: I6bc913440e60884203e21fea4375e94ea2ae1bbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061920
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>