[superproject] Roll cobalt-registry Add code font to .yaml.
cobalt-config try job fails for the change of removing legacy privacy
fields. This is because it checks the config not only for the existing
version of registries, but also the one before. So it fails the "check
parent config" task.

The purpose of this cl is to make sure both of the "current" and "parent" version of
registries does not contains legacy privacy fields.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/1062772
Privacy-Approval: Steve Fung <stevefung@google.com>
Original-Revision: 6b88db56becd54615aba573f13cf58c48e82b1a4
GitOrigin-RevId: 3665bf878d1d291efda9a5e6c323859bb21031ca
Roller-URL: https://ci.chromium.org/b/8745500875792275857
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id8a6c766859e0eb73251dd356fbad1e07a4d2676
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062994