[bazel_sdk] Fix bazel_test.py bug for --fuchsia_sdk_dir option.
Fix a subtle bug in the implementation of the `--fuchsia_sdk_dir`.
The location of the @fuchsia_sdk and @fuchsia_sdk_common
repositories was not overriden properly when running the
bazel test command, only when performing the queries that
compute the implicit dependencies for depfile generation.

This fixes the issue by correcting the way the BazelRepositoryMap
is initialized.

Change-Id: I0388ca2e99679ec6d7a4c37934bf9c8b44e1c337
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062859
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>