[build] Add the concept of compilation modes
This starts the replacement of the 'is_debug' GN arg and the use of
many fine-grained GN arguments with a set of 3 easy-to-specify set of
compilation modes (similar to Bazel):
- dbg
- fastbuild
- opt

The dbg and opt modes correspond to the defaults obtained today with
'is_debug = true' and 'is_debug = false'.

The new 'fastbuild' mode is a non-debug mode that uses a new
optimization level of 'minimal'.

Bug: b/339031006
Change-Id: I97ab801585d0f6fe01f0ccf895e63820f4532077
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042487
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>