[kernel] Teach the timer layer about boot time
This is the first change that introduces the notion of boot time to the
kernel. Boot time is a new timeline that includes the amount of time the
kernel has spent in a suspended state.

Mechanically, this change introduces several timer APIs that allow
higher level code to get the boot time and update the monotonic time
offset.

Note, however, that this change does not diverge the two timelines. In
other words, the monotonic offset and the boottime offset are kept
identical, so functionally this change should be a no-op. The point of
this CL is to ensure that the APIs have the shape we want and that the
memory order on the newly mutable mono_ticks_offset is accurate.

Bug: 341785588
Change-Id: I121f8ebfeb3a0e353c5e76bed5f10062d1d3e059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061012
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>