[bazel-sdk] do not collect resources for tools
Driver tools are binaries that get added to a driver package. They are
just cc binaries and get added to the bin/ directory. They do not bring
their own resources but we were checking for resources when we were
packaging which was causing a failure.

This CL adds a tests for tools and removes the check in the packing
logic for additional tool resources.

Change-Id: Ic8e38366f9e472d40db8af942a7d2f485cf3dee9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063274
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>