[build] Add optimization overrides to compilation_modes
This allows the overridding of options set by the compilation_mode,
starting with the 'optimize' GN arg.

This finishes the establishing of the mechanism for specifying
compilation modes, and allowing the args that are configured
by them to be overridden by the developer.

Bug: b/339031006

Change-Id: I79ab7b8d7fce43c529a7139924d3ea032c03906e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043374
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>