[display][coordinator] Remove async-watchdog.
The async-watchdog library provides a Watchdog class that
monitors the aliveness of a given async dispatcher by
posting periodical "update" tasks on the monitored dispatcher
and periodical "check" tasks on the Watchdog-owned loop.
In the display coordinator driver, it created an async::Loop
not managed by the driver framework, which is a discouraged
(and soon-to-be disallowed) behavior.

This change deletes the async-watchdog from the display
coordinator. Future changes will add other mechanisms to
avoid deadlocks within the display coordinator.

Test: no behavior change expected.
Test: no behavior change in the internal product on nelson.
Test: no behavior change in the internal product on VIM3.
Bug: 346419450
Bug: 329708151
Change-Id: I99c466802c7b2b4df920f9490cd74a937382e550
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063543
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>