[developer] Always use in-development level for protocol compatibility check
Previously, Fuchsia builds used `__Fuchsia_API_level__`, which will
never match `CURRENT_SUPPORTED_API_LEVEL` in platform builds. To fix
this, use the same build variable as used for host tool builds.

Rename FUCHSIA_API_LEVEL to NEXT_STABLE_API_LEVEL to avoid confusion
and prepare for `NEXT` when the in-development level will not be
numbered. This name is consistent with the value of the GN variable
`platform_version.in_development_api_level` from which it gets its
value and avoids needing to adjust any logic in the header.

Also ensure NEXT_STABLE_API_LEVEL is not less than
`CURRENT_SUPPORTED_API_LEVEL` because the logic does not allow this.

This fixes the naming issue described in the bug. This entire mechanism
will be replaced in the future.

Bug: 308808613
Change-Id: I4eb893b2cce8147c008bba277e77357759274368
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063701
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>