[bazel] pass compilation mode to refresh-compdb
The refresh-bazel-compdb command was not taking the compilation mode
into account when running the query which meant that all of the paths
were pointing to the fastbuild directories which do not exist in most
builds. This CL figures out what compilation mode you are using for the
normal build and mimics it.

Change-Id: I258bac052e1417cd2cc38e2cf688fdd86ebf860c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064437
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>