[kernel] Refactor TimerQueue in preparation for boottime
This CL refactors TimerQueue::Tick() to call a new TickInternal()
method that, in addition to the current timestamp and CPU that Tick
takes, also takes a reference to the timer_list_ to operate on.

This is a no-op today, but will be useful in the future when the
TimerList contains two different timer lists (one for the boottime
timeline and one for monotonic).

Because the two timelines technically operate on different time types
(zx_time_t and zx_boot_time_t), TickInternal is templated on this type.
Of course, given that all of these time types are typedefs of the same
underlying type, this won't matter today. However, this will matter in
the future when stronger time types are in use. Moreover, it provides
clarity to callers that the type of the `now` argument is not always
zx_time_t.

Bug: 341785588
Change-Id: Iebdfb2e0444f1bf9654ce6ce67b9ae02ddbea278
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060142
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>