[utest][core] Remove tests with debug dependencies
Removes a couple of tests that were using kernel shell commands for
functionality. These shell commands are not an API, subject to change,
and not always available depending on the build of the kernel.

The functionality these tests are attempting to validate is either
already tested as best as possible with in kernel tests, or is not
behavior that is guaranteed and predictable and therefore cannot have
a reliable test.

Bug: 340232040
Change-Id: I213421a362ae4e8a9b5ef71e33ceb692cec09177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064072
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>