[idk] Really build sdk_fuchsia_packages at PLATFORM
https://fxrev.dev/1026635 caused `sdk_fuchsia_package()` targets to be
`sdk_noop_atom()` targets when targeting PLATFORM. As a result, these
targets may not be built in a `:default` build, and some GN rules are
not even evaluated as they were for sub-builds. To avoid surprises and
restore build coverage for developers, restore the behavior prior to
that CL.

This CL effectively reverts the changes to this file that were made in
the CL above, but the implementation and comments are a bit different.

Also add logic to ensure that the SDK package manifest is reprocessed to
avoid manifest verification failures due to the change in how the change
in atom type of affected packages.

Bug: 310006516, 335204585
Test: Add `print("magma integration tests", current_toolchain)` to //src/graphics/magma/tests/integration/BUILD.gn and observe "//build/fidl:fidling" when `fx build` is run.
Change-Id: I48100d5625144c49ca154643ee83868193374168
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058658
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>