[bazel-sdk] Allow disabling package validation.
I'm soon going to increase the strictness of ABI compatibility checks,
and it'd be good to have an escape hatch in place if a product owner
needs to turn these checks off.

Turns out, assembly already has a flag for downgrading package
validation errors to warnings, so this CL simply plumbs that flag up to
the bazel level.

Theoretically, we could have a more granular check - either checking
just a subset of packages, or disabling just these stricter ABI revision
checks, but ideally this flag won't actually be used. If it turns out we
need more of a scalpel than a hacksaw, we can add one in the future.

Bug: 42082683

Change-Id: Ia28d7345e7d95ce55b6bf91051426cfd41797e77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063512
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>