[zircon][scheduler] Move the concept of "active" from mp -> Scheduler.
Prior to this change, the `mp` layer of the system was maintaining 3
different masks of per-cpu state in order to reflect 3 different
concepts.

1) The set of CPUs which are currently "online".
2) The set of CPUs/Schedulers which are currently "active".
3) The set of CPUs/Schedulers which are currently "idle".

This CL applies feedback from
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/1056093 to move
concept #2 out of the `mp` layer and into the `class Scheduler` layer.

This reflects the fact that "active" is really a Scheduler concept
(scheduler is active and accepting threads) and not so much a CPU
concept like "online" is ("online" CPUs are powered on and not
halted).

Change-Id: I0492b2927a2f3c30039b7bb40f9031c058893a9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064308
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adrian Danis <adanis@google.com>