[zircon][scheduler] Move the concept of "idle" from mp -> Scheduler.
Prior to this change, the `mp` layer of the system was maintaining 3
different masks of per-cpu state in order to reflect 3 different
concepts.

1) The set of CPUs which are currently "online".
2) The set of CPUs/Schedulers which are currently "active".
3) The set of CPUs/Schedulers which are currently "idle".

This CL applies feedback from
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/1056093 to move
concept #3 out of the `mp` layer and into the `class Scheduler` layer.

This reflects the fact that "idle/busy" is really a Scheduler concept
(scheduler last picked the idle/power thread to run instead of a
different thread to run) and not so much a CPU concept like "online"
is ("online" CPUs are powered-on and not halted).

Change-Id: Ifec96be280e3919ed6a47eb925949b0f92b6b2b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064309
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>