[display][coordinator] Replace display_mode_t with DisplayTiming.
This replaces the display_mode_t used in DisplayInfo for hardcoded
display timings, with the api-types-cpp DisplayTiming struct type.
This change also unifies the returned value type of Controller::
GetPanelConfig() and renamed it to GetDisplayTimings().

The only behavior change is that it's now valid to call SetDisplayMode()
on displays with a fixed display mode, as long as the display mode
in the FIDL call argument matches the fixed mode. The coordinator
will treat this as a no-op, so there'll be no difference in display
configs received by the display engine drivers.

on core.vim3:
Test: "display-tool squares" on vim3 with Amelin AML028-30MB-A1 (HDMI) works.
Test: "display-test --mode-set 1 1" on vim3 with Amelin AML028-30MB-A1 works.
Test: "display-tool squares" on vim3 with Amelin AML028-30MB-A1 works
      after device unplug and replug.
Test: "display-tool squares" on vim3 with Khadas TS050 (MIPI-DSI) works.
Test: "display-test --mode-set 1 0" on vim3 with Khadas TS050 works.

Test: "display-tool squares" works on core.nelson.

Bug: 42078780
Bug: 341156315
Change-Id: Ifce4d14474cef59c8b4c61f08443430c9a8ebed6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057126
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>