[reachability] Improve Component Exiting Flow
It was possible for the reachability eventloop to exit, which is most
frequently caused by losing connection to the netstack and the watcher
channels closing. Instead of using the ? operator throughout the `run`
body, replace with a call to std::process::exit so that the logs don't
have a reachability panic. Since std::process::exit doesn't output a
stack trace, the ambiguous source of `exit_with_fidl_error` has been
migrated to `exit_with_error` and takes an anyhow error that has a
string explanation of what exactly caused the eventloop to exit.

Change-Id: I83ec5c61d270a2922aa98e7227e64c7b7a2703a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048470
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>