[starnix] Test that futex_wait can be interrupted.
If a thread waiting on a futex receives a signal that triggers a signal
handler, and the signal disposition didn't set `SA_RESTART`, the futex
syscall should return with a `EINTR` error number.

This CL adds a test for that. The main thread sets up a signal handler
for `SIGUSR1` that doesn't do anything, spawns the second thread and
waits on a futex. The second thread blocks `SIGUSR1` on that thread, and
sends a process-directed `SIGUSR1` every 100ms.

The expected result is that the main thread is interrupted while waiting
for the futex.

Change-Id: If212014ac408ad8cd7b8fd021722c931e599be83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065248
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>