[netcfg] Prevent crashing from DHCPv6 client stops
Occasionally, a bunch of events will occur at the same time when trying
to start a DHCPv6 client. Netcfg uses multiple signals to determine when
the DHCPv6 client should be stopped. If those signals are handled at the
same time, netcfg does not know if the stop has already been handled and
will call it again.

Log a warning when the DHCPv6 client has already been stopped since this
is not always unreachable.

Bug: b/336787808
Change-Id: Ie0f17dc1e8e2caae53fc2bca09e7a50e35be31f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065703
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>