[netstack3] Implement SO_ORIGINAL_DST
Support the SO_ORIGINAL_DST socket option for TCP sockets, to allow
callers to retrieve the pre-NAT destination of a connected socket. This
option is valid for both IPv4 and IPv6 sockets, but for IPv6 sockets, it
will only return success if the socket is a dual-stack socket and the
connection's original destination was an IPv4 address.

Note that we inspect the TCP socket connection state and the connection
tracking state under the same lock to avoid a race condition.

Fixed: 338042280
Test: netstack-socket-integration-test
      netstack-filter-fidl-tests
Change-Id: I79f98bf1782e66e1ce46a3b6d6985e2d7ee4b3bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059293
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>