[brcmfmac] Fix SDIO teardown on Start failures
If initialization of the SDIO bus failed for any reason the bus object
would not be stored and the driver would fail to start. However, because
netdevice was initialized before the bus there could be netdev requests
in flight that would then be served by a non-existent bus. This would
then attempt to access a null pointer and crash.

Fix this by initializing netdevice after the bus has been successfully
initialized. Also do a complete shutdown of the device if bus
initialization fails to ensure that any resources that were initialized
are properly disposed of.

Bug: 344662647
Test: Manual testing with missing firmware files

Change-Id: I971a5bdd57567246554e36193eccb14702baf417
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065872
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>