[sensors] Loosen constraint to fix flakey test.
In retrospect, one of the constraints I put on the statistics in
the playback component integration test was not entirely
reasonable. The test has flaked a couple times on this constraint
and I believe it is safe to remove it. The other constraints should
be sufficient.

Change-Id: Ib57e48fabdf4d741536604d5d2031f7ac8967d2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065293
Commit-Queue: Matt Thiffault <mthiffault@google.com>
Reviewed-by: John Shamoon <johnshamoon@google.com>