[tee] Refactor the way API implementations are provided
This refactors the TEE Internal Core API implementations and adds some
extra safety to the bindings pattern. There are now 3 layers in the
implementation of TEE bindings:

- tee_internal::binding contains the bindgen generate Rust definitions
  of the TEE internal APIs
- api_impl::binding_stubss contains hand written mechanical forwarding
  stubs that implement the bindgen extern "C" functions by calling a
  (non-extern) Rust function in the api_impl crate.
- api_impl has our implementations of the entry points as regular
  Rust functions within the runtime.

The stubs layer exists so that we can get Rust type checking on our
implementations as the generated extern "C" functions do no type
checking on parameters and return values. This also adds new tests to
check that we are exposing the entry points that we expect to be.

This is similar to https://fxrev.dev/1048382 except that this changes
does not include the memory allocation implementations which need some
more work.

Bug: b/332956721
Change-Id: I5d9671489baf66ea99ea48c8c15cb85cd6603de0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065242
Reviewed-by: Eric Stone <ecstone@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>